Issue #10 Fix
Fixes Issue #10: https://github.com/bakercp/ofxClipper/issues/10

Changes in openframeworks ofPath class are the reason for this issue.  This code updates ofxClipper so it is compatible with the current state of ofPath.  Please be mindful that I am currently working off of a nightly build (of_v20160527_osx_release).