Allow clipping operations on open paths
ClipperLib can handle open paths on the subject, if the use_lines
preprocessing macro is defined.

If that is the case, we will by default use the PolyTree data structure
in the clip function, and we will also not close the paths if they are
added to the subjects vector.